ajout du test file + makefile compile avec un main comme usr la branche promptHistory
